Rollup merge:
2020.05.26:
* fixed: cpp: Linux x86_64 build fixes

2020.05.25:
* fixed: cpp: Linux x86_64 build fixes
* changed: cmake: Project.cmake: check construction at least one valid file path


git-svn-id: svn+ssh://svn.code.sf.net/p/tacklelib/tacklelib/trunk@392 ffccfebd-399a-4e7c-9cd2-5fb927d1fecb

+++ include/tacklelib/utility/memory.hpp
@@ -151,7 +151,7 @@ namespace utility
             check_buffer_guards();
 
             // minimum 16 bytes or 1% of allocation size for guard sections on the left and right, but not greater than `s_guard_max_len`
-            const size_t offset = (std::min)((std::max)(size / 100, 16U), _guard_max_len());
+            const size_t offset = (std::min)((std::max)(size / 100, size_t(16U)), _guard_max_len());
             const size_t size_extra = size ? (size + offset * 2) : 0;
 #else
             const size_t offset = 0;
